+++ braindecode/models/eeginception_mi.py
@@ -76,6 +76,7 @@ class EEGInceptionMI(EEGModuleMixin, nn.Module):
             in_channels=None,
             n_classes=None,
             input_window_s=None,
+            add_log_softmax=True,
     ):
         n_chans, n_outputs, input_window_seconds, = deprecated_args(
             self,
@@ -90,6 +91,7 @@ class EEGInceptionMI(EEGModuleMixin, nn.Module):
             n_times=n_times,
             input_window_seconds=input_window_seconds,
             sfreq=sfreq,
+            add_log_softmax=add_log_softmax,
         )
         del n_outputs, n_chans, chs_info, n_times, input_window_seconds, sfreq
         del in_channels, n_classes, input_window_s
@@ -165,13 +167,15 @@ class EEGInceptionMI(EEGModuleMixin, nn.Module):
 
         self.flat = nn.Flatten()
         self.fc = nn.Linear(
-            # in_features=self.input_window_samples * intermediate_in_channels,
             in_features=intermediate_in_channels,
             out_features=self.n_outputs,
             bias=True,
         )
 
-        self.softmax = nn.LogSoftmax(dim=1)
+        if self.add_log_softmax:
+            self.out_fun = nn.LogSoftmax(dim=1)
+        else:
+            self.out_fun = nn.Identity()
 
     def forward(
             self,
@@ -198,7 +202,7 @@ class EEGInceptionMI(EEGModuleMixin, nn.Module):
         out = self.ave_pooling(out)
         out = self.flat(out)
         out = self.fc(out)
-        return self.softmax(out)
+        return self.out_fun(out)
 
 
 class _InceptionModuleMI(nn.Module):