+++ braindecode/models/hybrid.py
@@ -9,9 +9,10 @@ from torch.nn import ConstantPad2d
 from .deep4 import Deep4Net
 from .util import to_dense_prediction_model
 from .shallow_fbcsp import ShallowFBCSPNet
+from .base import EEGModuleMixin, deprecated_args
 
 
-class HybridNet(nn.Module):
+class HybridNet(EEGModuleMixin, nn.Module):
     """Hybrid ConvNet model from Schirrmeister et al 2017.
 
     See [Schirrmeister2017]_ for details.
@@ -27,28 +28,46 @@ class HybridNet(nn.Module):
        Online: http://dx.doi.org/10.1002/hbm.23730
     """
 
-    def __init__(self, in_chans, n_classes, input_window_samples):
-        super(HybridNet, self).__init__()
+    def __init__(self, n_chans=None, n_outputs=None, n_times=None,
+                 in_chans=None, n_classes=None, input_window_samples=None,
+                 add_log_softmax=True):
+
+        n_chans, n_outputs, n_times = deprecated_args(
+            self,
+            ('in_chans', 'n_chans', in_chans, n_chans),
+            ('n_classes', 'n_outputs', n_classes, n_outputs),
+            ('input_window_samples', 'n_times', input_window_samples, n_times),
+        )
+        super().__init__(
+            n_outputs=n_outputs,
+            n_chans=n_chans,
+            n_times=n_times,
+            add_log_softmax=add_log_softmax,
+        )
+
         deep_model = Deep4Net(
-            in_chans,
-            n_classes,
+            n_chans=n_chans,
+            n_outputs=n_outputs,
             n_filters_time=20,
             n_filters_spat=30,
             n_filters_2=40,
             n_filters_3=50,
             n_filters_4=60,
-            input_window_samples=input_window_samples,
+            n_times=n_times,
             final_conv_length=2,
         )
         shallow_model = ShallowFBCSPNet(
-            in_chans,
-            n_classes,
-            input_window_samples=input_window_samples,
+            n_chans=n_chans,
+            n_outputs=n_outputs,
+            n_times=n_times,
             n_filters_time=30,
             n_filters_spat=40,
             filter_time_length=28,
             final_conv_length=29,
         )
+
+        del n_outputs, n_chans, n_times
+        del in_chans, n_classes, input_window_samples
 
         reduced_deep_model = nn.Sequential()
         for name, module in deep_model.named_children():
@@ -83,7 +102,7 @@ class HybridNet(nn.Module):
         self.reduced_deep_model = reduced_deep_model
         self.reduced_shallow_model = reduced_shallow_model
         self.final_conv = nn.Conv2d(
-            100, n_classes, kernel_size=(1, 1), stride=1
+            100, self.n_outputs, kernel_size=(1, 1), stride=1
         )
 
     def forward(self, x):
@@ -110,6 +129,9 @@ class HybridNet(nn.Module):
 
         merged_out = torch.cat((deep_out, shallow_out), dim=1)
         linear_out = self.final_conv(merged_out)
-        softmaxed = nn.LogSoftmax(dim=1)(linear_out)
-        squeezed = softmaxed.squeeze(3)
+        if self.add_log_softmax:
+            output = nn.LogSoftmax(dim=1)(linear_out)
+        else:
+            output = nn.Identity()(linear_out)
+        squeezed = output.squeeze(3)
         return squeezed